Fix travis perl build targets
Travis builds were failing to test on the specified versions of
Perl due to the default Ubuntu VM change from Trusty to Xenial.

While the main fix is to specify trusty as the Ubuntu VM to build on,
this commit updates the travis build config to additionally:

- remove Perl 5.8 from the build targets

Although it's still worthwhile to test against Perl 5.8 (there are
probably some users out there still running it somewhere), the
main use of Travis here for me is to test development branches
before releasing to cpan.  My plan is to roll off old build
targets as they're added to the list.  cpantesters tests through
5.8, so on release, it will test and catch any issues on older
Perls.

- remove Perl 5.32 from the build targets

5.32 is not available as a build target.  Until the binary is
available through Travis, again, it's okay to rely on cpantesters.

- add perlbrew-cache directory

https://docs.travis-ci.com/user/caching/

Additionally, this commit adds Test::NoBreakpoints to the cpanm
install list so the nobreaks test will run during testing.

Issue #6